move some test to accounts_test.go
